+++ src/resources/beta/models.ts
@@ -3,6 +3,7 @@
 import { APIResource } from '../../resource';
 import { isRequestOptions } from '../../core';
 import * as Core from '../../core';
+import * as BetaAPI from './beta';
 import { Page, type PageParams } from '../../pagination';
 
 export class Models extends APIResource {
@@ -11,9 +12,36 @@ export class Models extends APIResource {
    *
    * The Models API response can be used to determine information about a specific
    * model or resolve a model alias to a model ID.
+   *
+   * @example
+   * ```ts
+   * const betaModelInfo = await client.beta.models.retrieve(
+   *   'model_id',
+   * );
+   * ```
    */
-  retrieve(modelId: string, options?: Core.RequestOptions): Core.APIPromise<BetaModelInfo> {
-    return this._client.get(`/v1/models/${modelId}?beta=true`, options);
+  retrieve(
+    modelId: string,
+    params?: ModelRetrieveParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<BetaModelInfo>;
+  retrieve(modelId: string, options?: Core.RequestOptions): Core.APIPromise<BetaModelInfo>;
+  retrieve(
+    modelId: string,
+    params: ModelRetrieveParams | Core.RequestOptions = {},
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<BetaModelInfo> {
+    if (isRequestOptions(params)) {
+      return this.retrieve(modelId, {}, params);
+    }
+    const { betas } = params;
+    return this._client.get(`/v1/models/${modelId}?beta=true`, {
+      ...options,
+      headers: {
+        ...(betas?.toString() != null ? { 'anthropic-beta': betas?.toString() } : undefined),
+        ...options?.headers,
+      },
+    });
   }
 
   /**
@@ -21,20 +49,36 @@ export class Models extends APIResource {
    *
    * The Models API response can be used to determine which models are available for
    * use in the API. More recently released models are listed first.
+   *
+   * @example
+   * ```ts
+   * // Automatically fetches more pages as needed.
+   * for await (const betaModelInfo of client.beta.models.list()) {
+   *   // ...
+   * }
+   * ```
    */
   list(
-    query?: ModelListParams,
+    params?: ModelListParams,
     options?: Core.RequestOptions,
   ): Core.PagePromise<BetaModelInfosPage, BetaModelInfo>;
   list(options?: Core.RequestOptions): Core.PagePromise<BetaModelInfosPage, BetaModelInfo>;
   list(
-    query: ModelListParams | Core.RequestOptions = {},
+    params: ModelListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<BetaModelInfosPage, BetaModelInfo> {
-    if (isRequestOptions(query)) {
-      return this.list({}, query);
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
     }
-    return this._client.getAPIList('/v1/models?beta=true', BetaModelInfosPage, { query, ...options });
+    const { betas, ...query } = params;
+    return this._client.getAPIList('/v1/models?beta=true', BetaModelInfosPage, {
+      query,
+      ...options,
+      headers: {
+        ...(betas?.toString() != null ? { 'anthropic-beta': betas?.toString() } : undefined),
+        ...options?.headers,
+      },
+    });
   }
 }
 
@@ -65,7 +109,19 @@ export interface BetaModelInfo {
   type: 'model';
 }
 
-export interface ModelListParams extends PageParams {}
+export interface ModelRetrieveParams {
+  /**
+   * Optional header to specify the beta version(s) you want to use.
+   */
+  betas?: Array<BetaAPI.AnthropicBeta>;
+}
+
+export interface ModelListParams extends PageParams {
+  /**
+   * Header param: Optional header to specify the beta version(s) you want to use.
+   */
+  betas?: Array<BetaAPI.AnthropicBeta>;
+}
 
 Models.BetaModelInfosPage = BetaModelInfosPage;
 
@@ -73,6 +129,7 @@ export declare namespace Models {
   export {
     type BetaModelInfo as BetaModelInfo,
     BetaModelInfosPage as BetaModelInfosPage,
+    type ModelRetrieveParams as ModelRetrieveParams,
     type ModelListParams as ModelListParams,
   };
 }