+++ tests/api-resources/MessageStream.test.ts
@@ -32,6 +32,19 @@ async function* messageIterable(message: Message): AsyncGenerator<MessageStreamE
           }
         : content.type === 'thinking' ? { type: 'thinking', thinking: '', signature: '' }
         : content.type === 'redacted_thinking' ? { type: 'redacted_thinking', data: '' }
+        : content.type === 'server_tool_use' ?
+          {
+            type: 'server_tool_use',
+            id: 'toolu_01Up7oRoHeGvhded7n66nPzP',
+            name: 'web_search',
+            input: {},
+          }
+        : content.type === 'web_search_tool_result' ?
+          {
+            type: 'web_search_tool_result',
+            tool_use_id: 'toolu_01Up7oRoHeGvhded7n66nPzP',
+            content: [],
+          }
         : assertNever(content),
       index: idx,
     };
@@ -58,6 +71,10 @@ async function* messageIterable(message: Message): AsyncGenerator<MessageStreamE
       throw new Error('thinking not implemented yet');
     } else if (content.type === 'redacted_thinking') {
       throw new Error('redacted_thinking not implemented yet');
+    } else if (content.type === 'server_tool_use') {
+      throw new Error('server_tool_use not implemented yet');
+    } else if (content.type === 'web_search_tool_result') {
+      throw new Error('web_search_tool_result not implemented yet');
     } else {
       assertNever(content);
     }
@@ -70,7 +87,13 @@ async function* messageIterable(message: Message): AsyncGenerator<MessageStreamE
 
   yield {
     type: 'message_delta',
-    usage: { output_tokens: 6 },
+    usage: {
+      output_tokens: 6,
+      input_tokens: null,
+      cache_creation_input_tokens: null,
+      cache_read_input_tokens: null,
+      server_tool_use: null,
+    },
     // @ts-ignore
     delta: { stop_reason: message.stop_reason, stop_sequence: message.stop_sequence },
   };
@@ -160,6 +183,7 @@ describe('MessageStream class', () => {
           input_tokens: 10,
           cache_creation_input_tokens: null,
           cache_read_input_tokens: null,
+          server_tool_use: null,
         },
       }),
     );
@@ -220,22 +244,22 @@ describe('MessageStream class', () => {
         },
         {
           "args": [
-            "{"type":"message_start","message":{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message_start","message":{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
         {
           "args": [
             "{"type":"content_block_start","content_block":{"type":"text","text":"","citations":null},"index":0}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
         {
           "args": [
             "{"type":"content_block_delta","delta":{"type":"text_delta","text":"Hello"},"index":0}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
@@ -249,7 +273,7 @@ describe('MessageStream class', () => {
         {
           "args": [
             "{"type":"content_block_delta","delta":{"type":"text_delta","text":" ther"},"index":0}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello ther","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello ther","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
@@ -263,7 +287,7 @@ describe('MessageStream class', () => {
         {
           "args": [
             "{"type":"content_block_delta","delta":{"type":"text_delta","text":"e!"},"index":0}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
@@ -277,7 +301,7 @@ describe('MessageStream class', () => {
         {
           "args": [
             "{"type":"content_block_stop","index":0}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":null,"stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
@@ -289,27 +313,27 @@ describe('MessageStream class', () => {
         },
         {
           "args": [
-            "{"type":"message_delta","usage":{"output_tokens":6},"delta":{"stop_reason":"end_turn","stop_sequence":null}}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message_delta","usage":{"output_tokens":6,"input_tokens":null,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null},"delta":{"stop_reason":"end_turn","stop_sequence":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
         {
           "args": [
             "{"type":"message_stop"}",
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "streamEvent",
         },
         {
           "args": [
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "message",
         },
         {
           "args": [
-            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null}}",
+            "{"type":"message","id":"msg_01hhptzfxdaeehfxfv070yb6b8","role":"assistant","content":[{"type":"text","text":"Hello there!","citations":null}],"model":"claude-3-opus-20240229","stop_reason":"end_turn","stop_sequence":null,"usage":{"output_tokens":6,"input_tokens":10,"cache_creation_input_tokens":null,"cache_read_input_tokens":null,"server_tool_use":null}}",
           ],
           "type": "finalMessage",
         },
@@ -342,6 +366,7 @@ describe('MessageStream class', () => {
           "cache_read_input_tokens": null,
           "input_tokens": 10,
           "output_tokens": 6,
+          "server_tool_use": null,
         },
       }
     `);
@@ -372,6 +397,7 @@ describe('MessageStream class', () => {
           input_tokens: 10,
           cache_creation_input_tokens: null,
           cache_read_input_tokens: null,
+          server_tool_use: null,
         },
       }),
     );
@@ -399,7 +425,7 @@ describe('MessageStream class', () => {
     const stream = anthropic.messages.stream(
       {
         max_tokens: 1024,
-        model: 'claude-2.1',
+        model: 'claude-3-7-sonnet-20250219',
         messages: [{ role: 'user', content: 'Say hello there!' }],
       },
       { maxRetries: 0 },
@@ -424,7 +450,7 @@ describe('MessageStream class', () => {
     const stream = anthropic.messages.stream(
       {
         max_tokens: 1024,
-        model: 'claude-2.1',
+        model: 'claude-3-7-sonnet-20250219',
         messages: [{ role: 'user', content: 'Say hello there!' }],
       },
       { maxRetries: 0 },