+++ tests/api-resources/messages/messages.test.ts
@@ -64,6 +64,7 @@ describe('resource messages', () => {
           name: 'name',
           cache_control: { type: 'ephemeral' },
           description: 'Get the current weather in a given location',
+          type: 'custom',
         },
       ],
       top_k: 5,
@@ -120,6 +121,7 @@ describe('resource messages', () => {
           name: 'name',
           cache_control: { type: 'ephemeral' },
           description: 'Get the current weather in a given location',
+          type: 'custom',
         },
       ],
     });